Fix Ammunition.SupplyCost GraphQL type to Float
Static data contains fractional values (0.1, 1.2) that broke Int serialization on unitDetail queries.

+++ backend/src/graphql/schema.ts
@@ -232,7 +232,7 @@ export const schema = `
     ModelFileName: String
     HUDMultiplier: Float
     CriticMultiplier: Float
-    SupplyCost: Int
+    SupplyCost: Float
     ResupplyTime: Float
     Damage: Float
     StressDamage: Float
